requirements.txt

requests
hyperliquid-python-sdk
